Logging: add custom DETAIL & remove ERROR & CRITICAL levels
This is in response to discussion yesterday on the status of
the logging in practice, where it was agreed that error cases
are handled adequately by standard exceptions but that it would
be useful to have a further informational level.

Note: the API ref docs & possibly some other aspects of the
PR #34 still need to be amended to reflect these level changes.
This will be addressed in the next commit.