Tests: for _reset_log_emergence_level & _disable_logging
Note: this includes a name change for the old
'_reset_log_severity_level' function, to the above, since (despite
being intended as private) it was no longer accurately described,
either resetting the level which is in fact a mixture of severity &
verbosity, or disabling the logging which doesn't influence the level.
The term 'emergence' covers the various levels and disable capability.